requirements.txt
dj-config-url
# Django Core
Django==5.0.3
django-environ==0.11.2
python-decouple==3.8

# Database - ATUALIZADO para psycopg3
psycopg[binary]>=3.1.0
dj-database-url==2.1.0

# Supabase
supabase>=2.4.0
storage3>=0.7.3
gotrue>=2.4.1
postgrest>=0.16.0
realtime>=2.0.0
supafunc>=0.4.0

# Pillow para imagens
Pillow>=10.0.0

# APIs Externas
requests>=2.31.0

# ========== OTIMIZAÇÕES DE ESCALABILIDADE ==========

# Cache com Redis
redis==5.0.1
django-redis==5.4.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Rate Limiting
django-ratelimit==4.1.0

# Monitoring e Performance
django-debug-toolbar==4.2.0  # Apenas desenvolvimento

# ========== SISTEMA DE RECOMENDAÇÕES IA ==========

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# REST APIs
djangorestframework==3.14.0

# IA Premium (Google Gemini)
google-generativeai==0.3.2
